Add files via upload
Property Management System Version 1.0

This version supports reading saved files and creating tables.
It needs buttons such as insertion, deletion and storage.
And this only shows personal accounts. This program requires family accounts, family members tabs, and menus.

It's still in the development stage. A lot of modifications are needed.